Wave 1: Auto-fixable clippy fixes (A1-A37)
Major: Code Quality Tooling, Phase: Clippy Fix Wave 1
- A1: clone_on_copy (26 instances) - removed .clone() on Copy types
- A3: redundant_closure (5 instances) - .map(|x| f(x)) -> .map(f)
- A6: derivable_impls (3 instances) - manual impl Default -> #[derive(Default)]
- A8: single_match (1 instance) - match with single pattern -> if let
- A9: unnecessary_lazy_evaluations (2 instances) - .or_else(|| x) -> .or(x)
- A10: unnecessary_cast (2 instances) - removed same-type casts
- A11: needless_borrow (1 instance) - removed unnecessary &
- A12: useless_conversion (3 instances) - removed .into() to same type
- A13: question_mark (2 instances) - let Some(x) = y else { return None } -> let x = y?
- A14: needless_range_loop (5 instances) - use .iter_mut().enumerate()
- A15: collapsible_match (1 instance) - collapsed nested match
- A18: manual_contains (1 instance) - .iter().any() -> .contains()
- A19: new_without_default (1 instance) - added impl Default for ProbityStore
- A20: unwrap_or_default (1 instance) - .or_insert_with(Vec::new) -> .or_default()
- A23: zombie_processes (2 instances) - ensured wait() on all paths
- A24: unused_variables (1 instance) - prefixed with _
- A25: empty_line_after_doc_comment (1 instance) - removed empty line
- A27: explicit_auto_deref (1 instance) - &*x -> &x
- A33: writing_&Vec/PathBuf (2 instances) - use slice/Path refs
- A34: push_after_creation (1 instance) - vec![] + .push() -> vec![item]
- A36: replacing_text_with_itself (8 instances) - removed no-op replacements
- A37: borrowed_expression (1 instance) - removed unnecessary borrow

Warning count: 166 -> 48 (71% reduction)
Tests: 266 passed

opencode 1.14.28; vllm/qwen3.6-27b-AWQ-BF16-INT4

+++ p2p/core-engine/src/crypto_server/kem_mlkem.rs
@@ -24,7 +24,7 @@ pub fn mlkem_768_keypair() -> Result<(SignatureBytes, SignatureBytes), CryptoErr
 pub fn mlkem_768_encapsulate(public_key: &SignatureBytes) -> Result<(SignatureBytes, SignatureBytes), CryptoError> {
     let mut pk: ForetiasKemPubKey = unsafe { std::mem::zeroed() };
     pk.bytes[..public_key.len()].copy_from_slice(public_key);
-    pk.len = public_key.len() as usize;
+    pk.len = public_key.len();
     let mut ct: ForetiasKemCiphertext = unsafe { std::mem::zeroed() };
     let mut ss: [u8; 32] = [0u8; 32];
     let rc = unsafe {
@@ -38,10 +38,10 @@ pub fn mlkem_768_encapsulate(public_key: &SignatureBytes) -> Result<(SignatureBy
 pub fn mlkem_768_decapsulate(secret_key: &SignatureBytes, ciphertext: &SignatureBytes) -> Result<SignatureBytes, CryptoError> {
     let mut sk: ForetiasKemSecretKey = unsafe { std::mem::zeroed() };
     sk.bytes[..secret_key.len()].copy_from_slice(secret_key);
-    sk.len = secret_key.len() as usize;
+    sk.len = secret_key.len();
     let mut ct: ForetiasKemCiphertext = unsafe { std::mem::zeroed() };
     ct.bytes[..ciphertext.len()].copy_from_slice(ciphertext);
-    ct.len = ciphertext.len() as usize;
+    ct.len = ciphertext.len();
     let mut ss: [u8; 32] = [0u8; 32];
     let rc = unsafe {
         foretias_mlkem_768_decapsulate(&sk, &ct, ss.as_mut_ptr())
